Dynamic Versus Static Hedging
This program measures and compares the performance of static hedging (using the OLS model) versus dynamic hedging using the DCC-GARCH model. The approach involves cross-hedging by aligning the standard variable rate with a position in the sterling London Interbank Offered Rate (LIBOR) future. To account for the loan structure, a tailored volatility model is introduced as input for the hedge ratios. The DCC model outperforms, particularly under conditions of increased correlation and volatility.